Move namespace and use defined types
