+++ test/Altinn.Profile.Tests/IntegrationTests/API/Controllers/UsersControllerTests.cs
@@ -6,7 +6,6 @@ using System.Net;
 using System.Net.Http;
 using System.Net.Http.Headers;
 using System.Net.Http.Json;
-using System.Runtime.Intrinsics.X86;
 using System.Text;
 using System.Text.Json;
 using System.Text.Json.Serialization;
@@ -18,6 +17,7 @@ using Altinn.Profile.Core.Person.ContactPreferences;
 using Altinn.Profile.Core.User.ContactInfo;
 using Altinn.Profile.Core.User.ProfileSettings;
 using Altinn.Profile.Models;
+using Altinn.Profile.Tests.IntegrationTests.Mocks;
 using Altinn.Profile.Tests.IntegrationTests.Utils;
 using Altinn.Profile.Tests.Testdata;
 using Altinn.Register.Contracts;
